+++ CHANGELOG.md
@@ -48,6 +48,7 @@ _This changelog only relates to the `raffia` crate._
 - Changed AST of expression in Sass `@return`, `@debug`, `@warn`, `@error` at-rule and module config.
 - Added partial support of Sass division.
 - Added support of allowing keyframe block in Sass `@mixin` and `@include` at-rule.
+- Added support of allowing Sass at-rule in `@keyframes` at-rule.
 
 ## v0.2.0